Configurable output loggers
Logging output can now be sent to STDOUT without Logger, which is useful
with the Docker Logentries container as it expects JSON output without
any Logger headers. This includes severity and a timestamp in the JSON.

For non-Docker use, the Logger time and severity human-readable
formatting is a useful default.